requirements.txt
absl-py==2.3.1
cloudpickle==3.1.1
contourpy==1.3.2
cycler==0.12.1
Farama-Notifications==0.0.4
fonttools==4.60.1
gymnasium==1.2.1
imageio==2.37.0
imageio-ffmpeg==0.6.0
immutabledict==4.2.2
job-shop-lib==1.7.0
kiwisolver==1.4.9
matplotlib==3.10.7
networkx==3.4.2
numpy==2.2.6
ortools==9.12.4544
packaging==25.0
pandas==2.3.3
pillow==12.0.0
protobuf==5.29.5
psutil==7.1.2
pyarrow==20.0.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
simanneal==0.5.0
six==1.17.0
typing_extensions==4.15.0
tzdata==2025.2
